requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.46.0
funcy==2.0
gensim==4.3.2
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
kiwisolver==1.4.5
konlpy==0.6.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mysql-connector-python==8.2.0
numexpr==2.8.6
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
protobuf==4.21.12
pyLDAvis==3.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
threadpoolctl==3.2.0
tzdata==2023.3
zipp==3.17.0
